# Conflicts:
#	app/Contracts/Services/EventService/EventService.php

+++ app/Console/Commands/delDubleType.php
@@ -2,8 +2,14 @@
 
 namespace App\Console\Commands;
 
+use App\Models\Event;
+use App\Models\Place;
 use App\Models\Sight;
+use App\Models\Status;
+use App\Models\Timezone;
+use Carbon\Carbon;
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\DB;
 
 class delDubleType extends Command
 {
@@ -28,14 +34,31 @@ class delDubleType extends Command
      */
     public function handle()
     {
-        ini_set('memory_limit', '2048M');
         $bar = $this->output->createProgressBar(Sight::query()->count());
         Sight::query()->orderBy('id')->chunk(1000, function ($sight) use($bar) {
             $sight->each(function($sight) use($bar) {
-                $types = $sight->types->pluck('id')->toArray();
-                $sight->types()->detach($types);
-                $sight->types()->attach(array_unique($types));
-                $bar->advance();
+                DB::beginTransaction();
+                try {
+                    if (isset($sight->organization_id) && Event::where('organization_id', $sight->organization->id)->exists()) {
+                        $types = $sight->types->pluck('id')->toArray();
+                        $sight->types()->detach($types);
+                        $sight->types()->attach(array_unique($types));
+                        $bar->advance();
+                    } else {
+                        $status = Status::where('name', 'Отказ')->first();
+                        $statuses = $sight->statuses;
+                        foreach ($statuses as $status) {
+                            $sight->statuses()->updateExistingPivot($status["id"], [
+                                "last" => false
+                            ]);
+                        }
+                        $sight->statuses()->attach($status->id, ["last" => true]);
+                        $bar->advance();
+                    }
+                    DB::commit();
+                } catch (\Exception $e) {
+                    DB::rollBack();
+                }
             });
         });
         $bar->finish();
